+++ tests/FunctorSpec.hs
@@ -11,6 +11,7 @@
 {-# LANGUAGE UndecidableInstances #-}
 #if __GLASGOW_HASKELL__ >= 708
 {-# LANGUAGE EmptyCase #-}
+{-# LANGUAGE RoleAnnotations #-}
 #endif
 
 {-# OPTIONS_GHC -fno-warn-name-shadowing #-}
@@ -107,6 +108,9 @@ data IntHashFun a b
 
 data Empty1 a
 data Empty2 a
+#if __GLASGOW_HASKELL__ >= 708
+type role Empty2 nominal
+#endif
 
 -- Data families